commented steps and refactor helper functions
