+++ src/components/big-calendar/big.calendar.container.jsx
@@ -2,24 +2,32 @@ import React, { useEffect, useState } from "react";
 // Import Component
 import BigCalendar from "./big.calendar.component";
 // import helpers
-import {
-  createCalendarDateObject,
-  createCalendarEventArray
-} from "../../helpers/big-calendar-helper";
+import { createCalendarEventArray } from "../../helpers/big-calendar-helper";
+import { sortPlantsByWateringDate } from "../../helpers/plant-helper";
+import { createDateArray, createDateObject } from "../../helpers/date-helpers";
 
-const BigCalendarContainer = ({ sortedPlants }) => {
+const BigCalendarContainer = ({ plants }) => {
+  // 1) sort the plants by water_after
+  const [sortedPlants] = useState(sortPlantsByWateringDate(plants));
   const [events, setEvents] = useState(null);
   useEffect(() => {
     (async () => {
       try {
-        // create event data structure for BigCalendar
-        const dates = await createCalendarDateObject(sortedPlants);
+        // 2) create an array of all the dates
+        const dateArray = await createDateArray(plants);
+
+        // 3) create a date object where the water_after is the key
+        //    and and array of dates are the values
+        const dates = await createDateObject(dateArray);
+
+        // 4) create an array with all the watering events, using the sortedPlants
+        //    and date object
         setEvents(await createCalendarEventArray(sortedPlants, dates));
       } catch (error) {
         console.log("couldnt create calendar event structure:", error);
       }
     })();
-  }, [sortedPlants]);
+  }, [plants, sortedPlants]);
   return (
     <div>
       {events ? (
